chore: extract performAddToPlanner() to be common operations
