+++ src/common/commonOperations.js
@@ -1,6 +1,8 @@
 import axios from 'axios';
 import ReactGA from 'react-ga';
 import i18n from 'i18next';
+import { isSpecialLectureCourse } from '../utils/courseUtils';
+import { getCourseOfItem } from '../utils/itemUtils';
 
 export const performSearchCourses = (option, limit, beforeRequest, afterResponse) => {
   if (
@@ -188,6 +190,80 @@ export const performDeleteFromCart = (lecture, user, fromString, beforeRequest, 
     action: 'Deleted Lecture from Cart',
     label: `Lecture : ${lecture.id} / From : ${fromString}`,
   });
+};
+
+export const performAddToPlanner = (
+  course,
+  year,
+  semester,
+  selectedPlanner,
+  user,
+  fromString,
+  beforeRequest,
+  afterResponse,
+) => {
+  const duplicateFutureItems = selectedPlanner.future_items.filter(
+    (fi) =>
+      !fi.is_excluded &&
+      !isSpecialLectureCourse(getCourseOfItem(fi)) &&
+      getCourseOfItem(fi).id === course.id,
+  );
+  if (duplicateFutureItems.length > 0) {
+    // eslint-disable-next-line no-alert
+    alert('동일한 과목의 수강 예정이 이미 추가되어 있습니다.');
+    return;
+  }
+
+  const duplicateTakenItems = selectedPlanner.taken_items.filter(
+    (ti) =>
+      !ti.is_excluded &&
+      !isSpecialLectureCourse(getCourseOfItem(ti)) &&
+      getCourseOfItem(ti).id === course.id,
+  );
+  if (duplicateTakenItems.length > 0) {
+    // eslint-disable-next-line no-alert
+    const userConfirmed = window.confirm(
+      '동일한 과목의 수강 기록이 플래너에 이미 추가되어 있습니다. 정말 추가하시겠습니까? 이전에 수강한 과목은 제외 처리됩니다.',
+    );
+    if (!userConfirmed) {
+      return;
+    }
+  }
+
+  beforeRequest();
+
+  if (!user) {
+    const id = this._createRandomItemId();
+    const item = {
+      id: id,
+      item_type: 'FUTURE',
+      is_excluded: false,
+      course: course,
+      year: year,
+      semester: semester,
+    };
+    afterResponse(item, duplicateTakenItems);
+  } else {
+    axios
+      .post(
+        `/api/users/${user.id}/planners/${selectedPlanner.id}/add-future-item`,
+        {
+          course: course.id,
+          year: year,
+          semester: semester,
+        },
+        {
+          metadata: {
+            gaCategory: 'Planner',
+            gaVariable: 'POST Update / Instance',
+          },
+        },
+      )
+      .then((response) => {
+        afterResponse(response.data, duplicateTakenItems);
+      })
+      .catch((error) => {});
+  }
 };
 
 export const performSubmitReview = (
